Fix bug in defaults for Dynesty
Previously, the dynesty kwarg for the number of live points was set to
be `npoints`, and so `npoints=10000` was passed to
`dynesty.NestedSampler`. This, however, is not the kwarg requried by
`NestedSampler`, but it instead should be `nlive`.  Unfortunately, this
error is silenced because the kwargs are just lost in dynesty.

- Fixes nlive kwarg
- Reduces default number from 10000 to 250 (as this was what was
  actually used before)
- Adds a verbose flag controlling the `run_nested` argument
  `print_progress`