Simplifies the logic of how parameters are known-about from the sampler
The samplers and prior access the parameters via
`likelihood.waveform_generator.parameters`. But, this is the only way
that the waveform_generator was accessed outside of the likelihood. We
simplify this by insisting that the Likelihood has the attribute
parameters (which is set, in this case, from the waveform_generator).

- Add attribute parameters to likelihood
- Remove all look-ups of waveform_generator outside of the likelihood.
  This means the waveform_generator is a likelihood-specific thing!