Add explicit methods
This adds methods to __dir__ so that plot_corner and other methods can
be tab completed (in ipython). Unfortunately, this is done by setting
these directly. While this isn't an ideal solution - there is no obvious
way to automate the process as methods to looks up all methods
themselves use the __dir__ attribute causing a recursion error.

Closes #111