Small simplification of waveform generator / likelihood interaction
This allows the waveform generator to be given None for the sampling
frequency and duration (in this state, it therefore can't be used to
generate an injection). Then, when passed to the likelihood, a check if
performed to see if those attributes are set. If they are None, they are
set from the data, and if they disagree with the data, an error is raised.