Unify the creation of the outdir
Previously, outdir was checked and if-not-existing created in a few
different places in two different ways. This (a) makes them all use the
same method (namely the utils function) and (b) places a
check-and-creation call before every tupak output. This will ensure that
changes elsewhere in the code won't cause bugs in the future. Also, if
you choose not to save anything, it won't just make an empty directory.