Fix bug in combining evidences
If `use_ratio` was None, the `log_evidence` was recalculated correctly,
but the log_bayes_factor was not updated. This meant the
log_bayes_factor from the first result file was stored causing
inconsistencies (the log_ev - log_ev_err != log_bf). This fixes this
issue and makes the interface more straight forward, always using the
log_evidence directly